Mon, Jun 26, 2023, 4:49 PM -04:00

+++ week2/lecture/scores.c
@@ -2,10 +2,10 @@
 
 int main(void)
 {
-    float score1 = 72;
-    float score2 = 73;
-    float score3 = 33;
+    int score1 = 72;
+    int score2 = 73;
+    int score3 = 33;
 
-    printf("Average: %f\n", (score1 + score2 + score3) / 3);
+    printf("Average: %f\n", (score1 + score2 + score3) / 3.0);
 
 }
